Be more robust against unknown entry types
Right now when an unknown entry type is encountered, we can end up
appending a NULL value in the middle of the entry list, which might
later result in a segfault in the client dereferencing NULL when
iterating the list.
Be more robust to that error, so that the returned entry list is always
composed of valid elements.

https://bugzilla.redhat.com/show_bug.cgi?id=872884

https://bugzilla.gnome.org/show_bug.cgi?id=693353